add imagepicker to Add tip screen
