+++ App.js
@@ -1,7 +1,6 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import AddTipScreen from "./app/screens/AddTipScreen";
-import SelectImage from "./app/components/AddTipForm/SelectImage";
 
 export default function App() {
-  return <SelectImage />;
+  return <AddTipScreen />;
 }
