--- app/components/AddTipForm/SelectImage.js
@@ -1,24 +1,0 @@
-import React, { useState } from "react";
-import Screen from "../Screen";
-import ImageInputList from "./ImageInputList";
-
-export default function SelectImage() {
-  const [imageUris, setImageUris] = useState([]);
-
-  const handleAdd = (uri) => {
-    setImageUris([...imageUris, uri]);
-  };
-
-  const handleRemove = (uri) => {
-    setImageUris(imageUris.filter((imageUri) => imageUri !== uri));
-  };
-  return (
-    <Screen>
-      <ImageInputList
-        imageUris={imageUris}
-        onAddImage={handleAdd}
-        onRemoveImage={handleRemove}
-      />
-    </Screen>
-  );
-}
